Rename material to pbr
